Codegen for openapi v216
