+++ lib/stripe/object_types.rb
@@ -97,11 +97,11 @@ module Stripe
         Subscription::OBJECT_NAME => Subscription,
         SubscriptionItem::OBJECT_NAME => SubscriptionItem,
         SubscriptionSchedule::OBJECT_NAME => SubscriptionSchedule,
-        TaxCalculation::OBJECT_NAME => TaxCalculation,
+        Tax::Calculation::OBJECT_NAME => Tax::Calculation,
+        Tax::Transaction::OBJECT_NAME => Tax::Transaction,
         TaxCode::OBJECT_NAME => TaxCode,
         TaxId::OBJECT_NAME => TaxId,
         TaxRate::OBJECT_NAME => TaxRate,
-        TaxTransaction::OBJECT_NAME => TaxTransaction,
         Terminal::Configuration::OBJECT_NAME => Terminal::Configuration,
         Terminal::ConnectionToken::OBJECT_NAME => Terminal::ConnectionToken,
         Terminal::Location::OBJECT_NAME => Terminal::Location,
