+++ lib/stripe/resources.rb
@@ -84,11 +84,11 @@ require "stripe/resources/source_transaction"
 require "stripe/resources/subscription"
 require "stripe/resources/subscription_item"
 require "stripe/resources/subscription_schedule"
-require "stripe/resources/tax_calculation"
+require "stripe/resources/tax/calculation"
+require "stripe/resources/tax/transaction"
 require "stripe/resources/tax_code"
 require "stripe/resources/tax_id"
 require "stripe/resources/tax_rate"
-require "stripe/resources/tax_transaction"
 require "stripe/resources/terminal/configuration"
 require "stripe/resources/terminal/connection_token"
 require "stripe/resources/terminal/location"
